Move compatibility extenstions to the separate directory
