+++ lib/anycable/rails/compatibility/ext/channel.rb
