+++ lib/anycable/rails/compatibility/ext/remote_connection.rb
